[Enhancement #216] Support updating referenced lists containing multilingual strings in the RDF4J driver.
